+++ src/main/java/compiler/nodes/NodeParameters.java
@@ -10,15 +10,21 @@ public class NodeParameters {
     
     // String is the var name, TokenType the type naturally
     private LinkedHashMap<String, Token> variables = new LinkedHashMap<>(); 
+    private ArrayList<Boolean> mutable = new ArrayList<>();
 
     public NodeParameters() {}
     
-    public void addVariable(String name, Token type) {
+    public void addVariable(String name, Token type, boolean isMutable) {
         this.variables.put(name, type);
+        this.mutable.add(isMutable);
     }
 
     public LinkedHashMap<String, Token> getVariables() {
         return this.variables;
+    }
+
+    public Boolean isMutable(Integer i) {
+        return mutable.get(i);
     }
 
     @Override