+++ CHANGELOG.md
@@ -45,4 +45,4 @@
 
 - Make sure something is actually returned for strings and integers (i.e. non void return methods)
 - Add in tests for UnaryExpression type
-- Support specifying variables as mut in function calls
+- Support specifying variables as mut in function calls